# openapi.yaml
openapi: 3.0.3
info:
  title: Node-Learn API
  version: 1.0.0
  description: Consolidated OpenAPI spec for the API Gateway and microservices.
  license:
    name: UNLICENSED
    url: https://example.com/unlicensed
servers:
  - url: /api
    description: API gateway base path
security: []
paths:
...
  /auth/google:
    get:
      tags: [auth]
      summary: Start Google OAuth flow
      operationId: startGoogleOAuth
      responses:
        '200':
          description: OAuth flow started
        '302':
          description: Redirect to Google OAuth
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/google/callback:
    get:
      tags: [auth]
      summary: Google OAuth callback
      operationId: handleGoogleOAuthCallback
      responses:
        '200':
          description: OAuth callback handled
        '302':
          description: Redirect to frontend with token
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/profile:
    get:
      tags: [auth]
      summary: Get current user profile
      operationId: getAuthProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /quiz/stats/record:
    post:
      tags: [quiz]
      summary: Record an answer (alias)
      operationId: recordQuizStats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordAnswer'
      responses:
        '200':
          description: Record result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordAnswerResult'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /stats/user/{userId}:
    get:
      tags: [stats]
      summary: Get user statistics
      operationId: getUserStats
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '400':
          $ref: '#/components/responses/BadRequest'
  /stats/user/{userId}/wrong-questions:
    get:
      tags: [stats]
      summary: Get user wrong questions
      operationId: getUserWrongQuestions
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Wrong questions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /stats/record:
    post:
      tags: [stats]
      summary: Record answer and achievements
      operationId: recordStats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordAnswer'
      responses:
        '200':
          description: Record result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordAnswerResult'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /achievements:
    get:
      tags: [achievements]
      summary: List achievements
      operationId: listAchievements
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: Achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Achievement'
        '400':
          $ref: '#/components/responses/BadRequest'
  /achievements/user/{userId}:
    get:
      tags: [achievements]
      summary: Get achievements for a user
      operationId: getAchievementsForUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserAchievement'
        '400':
          $ref: '#/components/responses/BadRequest'
  /achievements/check:
    post:
      tags: [achievements]
      summary: Check and award achievements
      operationId: checkAchievements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AchievementCheck'
      responses:
        '200':
          description: Awarded achievements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AchievementAwardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /achievements/{id}:
    get:
      tags: [achievements]
      summary: Get achievement by id
      operationId: getAchievementById
      parameters:
        - $ref: '#/components/parameters/AchievementIdPath'
      responses:
        '200':
          description: Achievement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /leaderboard/update:
    post:
      tags: [leaderboard]
      summary: Update leaderboard
      operationId: updateLeaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaderboardUpdate'
      responses:
        '200':
          description: Update result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
  /leaderboard/mode/{modeId}:
    get:
      tags: [leaderboard]
      summary: Get leaderboard for a mode
      operationId: getLeaderboardByMode
      parameters:
        - $ref: '#/components/parameters/ModeIdPath'
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
  /leaderboard/user/{userId}:
    get:
      tags: [leaderboard]
      summary: Get leaderboard position for a user
      operationId: getLeaderboardForUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Leaderboard entry
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: integer
    ModeIdPath:
      name: modeId
      in: path
      required: true
      schema:
        type: string
    AchievementIdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    UserIdQuery:
      name: userId
      in: query
      schema:
        type: string
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    RecordAnswer:
      type: object
      required: [userId, questionId, selectedChoiceId, quizModeId, isCorrect]
      properties:
        userId:
          type: string
        questionId:
          type: integer
        selectedChoiceId:
          type: integer
        quizModeId:
          type: integer
        isCorrect:
          type: boolean
    RecordAnswerResult:
      type: object
      properties:
        success:
          type: boolean
        awardedAchievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
        leaderboardUpdated:
          type: boolean
    Achievement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        sprite_col:
          type: integer
        sprite_row:
          type: integer
        created_at:
          type: string
          nullable: true
    AchievementProgress:
      type: object
      properties:
        current:
          type: integer
        total:
          type: integer
        label:
          type: string
    UserAchievement:
      allOf:
        - $ref: '#/components/schemas/Achievement'
        - type: object
          properties:
            unlocked:
              type: boolean
            unlocked_at:
              type: string
              nullable: true
            progress:
              $ref: '#/components/schemas/AchievementProgress'
    AchievementCheck:
      type: object
      properties:
        userId:
          oneOf:
            - type: string
            - type: integer
        questionId:
          type: integer
        isCorrect:
          type: boolean
        questionType:
          type: string
        practical:
          type: boolean
        difficulty:
          type: string
          enum: [easy, medium, hard]
    AchievementAwardResponse:
      type: object
      properties:
        awardedAchievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
        awarded:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
    UserStatsOverview:
      type: object
      properties:
        totalCorrect:
          type: integer
        totalIncorrect:
          type: integer
        totalAttempts:
          type: integer
        overallAccuracy:
          type: string
    UserStatsQuestion:
      type: object
      properties:
        id:
          type: integer
        question_id:
          type: integer
        question:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        correct_count:
          type: integer
        incorrect_count:
          type: integer
        accuracy_percentage:
          type: string
    UserStats:
      type: object
      properties:
        overall:
          $ref: '#/components/schemas/UserStatsOverview'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/UserStatsQuestion'
    LeaderboardUpdate:
      type: object
      required: [modeId, userId, correctAnswers, totalQuestions, correctAnswersCount, username]
      properties:
        modeId:
          type: string
        userId:
          type: integer
        correctAnswers:
          type: integer
        totalQuestions:
          type: integer
        correctAnswersCount:
          type: integer
        username:
          type: string
    LeaderboardEntry:
      type: object
      properties:
        quiz_mode_id:
          type: string
        position:
          type: integer
        user_id:
          type: integer
        correct_answers:
          type: integer
        total_questions:
          type: integer
        correct_answers_count:
          type: integer
        achieved_at:
          type: string
          format: date-time
        username:
          type: string
          nullable: true
    User:
      type: object
      properties:
        id:
          type: integer
        googleId:
          type: string
        email:
          type: string
        name:
          type: string
        picture:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time